Uppercase PNG extensions cause Error on upload
I'm working on a modified code and i pass original filename to jic.upload()
Problem is if PNG extension is uppercase the line 70 fails to set mimetype and tries to process a png image as a jpeg. As a result it fires the error below.
Uncaught InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded.
So i modified line 70 to lowercase extension.